Factor out json_importer.annotations.
